refactor: split whoiswrapper
