+++ app/ts/common/lookup.ts
@@ -1,0 +1,118 @@
+import psl from 'psl';
+import puny from 'punycode';
+import uts46 from 'idna-uts46';
+import whois from 'whois';
+import debugModule from 'debug';
+import { load, Settings } from './settings';
+
+const debug = debugModule('common.whoisWrapper');
+let settings: Settings = load();
+
+const lookupPromise = (...args: unknown[]): Promise<string> => {
+  return new Promise((resolve, reject) => {
+    (whois as any).lookup(...args, (err: unknown, data: string) => {
+      if (err) return reject(err);
+      resolve(data);
+      return undefined;
+    });
+  });
+};
+
+export async function lookup(domain: string, options = getWhoisOptions()): Promise<string> {
+  const {
+    'lookup.conversion': conversion,
+    'lookup.general': general,
+  } = settings;
+  let domainResults: string;
+
+  try {
+    domain = conversion.enabled ? convertDomain(domain) : domain;
+    if (general.psl) {
+      const clean = psl.get(domain);
+      domain = clean ? clean.replace(/((\*\.)*)/g, '') : domain;
+    }
+
+    debug(`Looking up for ${domain}`);
+    domainResults = await lookupPromise(domain, options);
+  } catch (e) {
+    domainResults = `Whois lookup error, ${e}`;
+  }
+
+  return domainResults;
+}
+
+export function convertDomain(domain: string, mode?: string): string {
+  const {
+    'lookup.conversion': conversion,
+  } = settings;
+
+  mode = mode || conversion.algorithm;
+
+  switch (mode) {
+    case 'punycode':
+      return puny.toASCII(domain);
+    case 'uts46':
+      return uts46.toAscii(domain);
+    case 'uts46-transitional':
+      return uts46.toAscii(domain, {
+        transitional: true,
+      });
+    case 'ascii':
+      return domain.replace(/[^\x00-\x7F]/g, "");
+    default:
+      return domain;
+  }
+}
+
+export function getWhoisOptions(): Record<string, unknown> {
+  const {
+    'lookup.general': general,
+  } = settings;
+
+  const options: Record<string, unknown> = {},
+    follow = 'follow',
+    timeout = 'timeout';
+
+  options.server = general.server;
+  options.follow = getWhoisParameters(follow);
+  options.timeout = getWhoisParameters(timeout);
+  options.verbose = general.verbose;
+
+  return options;
+}
+
+function getWhoisParameters(parameter: string): number | undefined {
+  const {
+    'lookup.randomize.follow': follow,
+    'lookup.randomize.timeout': timeout,
+    'lookup.randomize.timeBetween': timeBetween,
+    'lookup.general': general,
+  } = settings;
+
+  switch (parameter) {
+    case 'follow':
+      debug(`Follow depth, 'random': ${follow.randomize}, 'maximum': ${follow.maximumDepth}, 'minimum': ${follow.minimumDepth}, 'default': ${general.follow}`);
+      return (follow.randomize ? getRandomInt(follow.minimumDepth, follow.maximumDepth) : general.follow);
+    case 'timeout':
+      debug(`Timeout, 'random': ${timeout.randomize}, 'maximum': ${timeout.maximum}, 'minimum': ${timeout.minimum}, 'default': ${general.timeout}`);
+      return (timeout.randomize ? getRandomInt(timeout.minimum, timeout.maximum) : general.timeout);
+    case 'timebetween':
+      debug(`Timebetween, 'random': ${timeBetween.randomize}, 'maximum': ${timeBetween.maximum}, 'minimum': ${timeBetween.minimum}, 'default': ${general.timeBetween}`);
+      return (timeBetween.randomize ? getRandomInt(timeBetween.minimum, timeBetween.maximum) : general.timeBetween);
+    default:
+      return undefined;
+  }
+}
+
+function getRandomInt(min: number, max: number): number {
+  min = Math.floor(min);
+  max = Math.floor(max);
+  if (min > max) [min, max] = [max, min];
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+export default {
+  lookup,
+  convertDomain,
+  getWhoisOptions,
+};
