+++ app/ts/renderer/bwa/fileinput.ts
@@ -1,6 +1,5 @@
 
 /** global: settings */
-import * as whois from '../../common/whoiswrapper';
 import * as conversions from '../../common/conversions';
 import fs from 'fs';
 import Papa from 'papaparse';
